% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download dataset}
\usage{
download_data(
  database = NULL,
  url_database = NULL,
  file_format = NULL,
  data_name = NULL
)
}
\arguments{
\item{database}{}

\item{url_database}{}

\item{file_format}{}

\item{data_name}{}
}
\description{
Download dataset
}
